fix: Result.kt - 수익률 자료형 수정

+++ src/main/kotlin/lotto/domain/Result.kt
@@ -37,7 +37,7 @@ object Result {
         return setA.intersect(setB)
     }
 
-    fun calculateRatio(price: Int, result: List<Int>): Float {
+    fun calculateRatio(price: Int, result: List<Int>): Double {
         val reward = listOf<Int>(2000000000, 30000000, 1500000, 50000, 5000)
         var income = 0
 
@@ -45,7 +45,7 @@ object Result {
             income += reward[i]*result[i]
         }
 
-        val ratio:Float = (income.toFloat())/(price.toFloat())*100
+        val ratio:Double = (income.toDouble())/(price.toDouble())*100
 
         return ratio
     }
